Add files via upload
This is my project for US Medical insurance cost data. 
The goal of this analysis is to show the average insurance cost reagardign the groups from each feature (age, sex, smoking habits, children and region). 
Also I created a function which takes as input two of the features, and two of the values and return the average cost, so it can be used for comparison reasons of how different features impacts the insurance cost.